requirements.txt
langchain==0.3.26
langchain-community==0.3.26

flask==3.0.3

sentence-transformers==2.7.0
transformers==4.41.2

pypdf==4.2.0
python-dotenv==1.0.1

pinecone-client==3.2.2
langchain-pinecone==0.0.1

langchain-openai==0.1.7

faiss-cpu==1.8.0